docs/ref/databases: rephrase persistent connections section
At first glance, I thought this section (and the whole page)
referred to requests to the database, hinting that Django would have
recreated connections to the database after each query.

In reality, Django manages that only in the context of `HttpRequest`
calls. Hence, this commit clarifies this behavior in the documentation.